Add support for the new switcher dynamic features
When a favorite app is closed with the switcher visible, dynamically add
the corresponding launcher. There is no way to specify the index of the
launcher, so it is appended to the end of the list.

Also dynamically remove the launcher if the app starts running when the
switcher is visible.